requirements.txt
streamlit
pytube
openai-whisper
openai
python-dotenv